Remove Virtual Directories from documentation
I don't speak German, French, Italian and Japanese.
Here is what I did for these languages: I removed the
virtual-directories.md file. Then I searched for
'virtual' in the language directory (de, fr, etc.) to
find references to the .md file. Here, I found the
translation of virtual into the target language
(virtuelle, virtuel, etc.) and did a second seach for
that word in the language directory. This way I think
I have removed most (if not all) references to it in all
languages. I'm not so sure about Japanese :)

The FAQs on some languages (de, fr) still contain
a sentence about uninstalling Zettlr that mentions the
remains of a virtual directory. I changed the wording a bit
in the English version but did not touch the others. I don't
think it is a big problem anyway.