Update yaml-frontmatter.md
Added a sentence to the lang variable description to make it clear it covers more than just the language of one's references.  Also added a sentence to the nocite variable section to remind people to quote their value as @ symbols have special meaning in YAML frontmatter because I saw that had tripped someone up in another bug report.