Create moving.md
A brief guide for moving to a new computer or OS install. I wasn't quite sure where to place this. It's too big to fit neatly alongside the parsimonious answers in the FAQ document. It doesn't seem to be a subtopic of any of the existing docs. The other documents most like it are in the "Getting Started" section, but this is obviously aimed at folks who've been using Zettlr for a while (and perhaps those who are hesitant to commit until they have a sense of how portable their data will be). "References" seemed like the best fit, but it still feels different than the other documents there. If this doc gets accepted feel free to place it wherever it makes the most sense.