Update citations.md
I suggest updating the link to Better Bibtex. The current link points to the download page of the plugin .xpi file on GitHub. However, the Better Bibtex project has a user-friendly install guide that I suggest pointing to instead. This change makes sense to me because the link text in the Zettlr doc promises to help users install Better Bibtex. Linking to the Better Bibtex install guide (which, in turn, has a prominent link to the .xpi download page) is more user-friendly for folks who may be less familiar with Zotero or its plugin system.

For example, I'm about to try to have a class of undergrads use Zettlr and Zotero together. I plan to walk through the steps to connect the two in class, but many will try it beforehand based on the links they find in the syllabus. They will never have used Zotero or Zotero plugins before and will be completely mystified by the link in the Zettlr doc to a the Better Bibtex .xpi file.

If, instead, they clicked on the link in the Zettlr doc and were taken to a guide on how to install Better Bibtex and explains the nature of the .xpi file — then, I think things would be simpler for them.